requirements.txt
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
PyDrive==1.3.1
python-telegram-bot==11.1.0
requests==2.22.0
urllib3==1.25.3
vk-api==11.4.0
python-telegram-bot==11.1.0
